pyproject.toml
[project]
name = "jopka-kirill"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill Chekmenev",email = "chekmenev2004@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "power-grid-model (>=1.12.27,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "streamlit (>=1.49.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "windpowerlib (>=0.2.2,<0.3.0)",
    "streamlit-image-coordinates (>=0.4.0,<0.5.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pvlib (>=0.13.0,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
